refactor(deployment): move dev-wheel build lifecycle into wheel_build
Relocate the dev-wheel build, per-process cache, and staging-copy helpers
(_build_dev_wheel_cached, _copy_local_framework_for_override,
_reset_wheel_build_cache and their cache state) out of compose_generator
into wheel_build, which already owned the manifest and artifact-path
helpers. compose_generator re-exports the copy and cache-reset helpers so
existing import and patch call sites keep resolving against it, and retains
_stage_dev_wheel_for_context, the render-flow staging decision.

No behavior change; deployment tests updated only to patch the relocated
helpers in their new module namespace.
